+++ core/initializer/test_execution/test_executor/test_executor.go
@@ -245,12 +245,15 @@ func RunTest(
 		return false, stacktrace.Propagate(err, "An error occurred waiting for the exit of the Kurtosis API container: %v", err)
 	}
 
+	// At this point, we may be printing the logs of a stopped test suite container, or we may be printing the logs of
+	//  still-running container that's exceeded the hard test timeout. Regardless, we want to print these so the user
+	//  gets more information about what's going on, and the user will learn the exact error below
+	banner_printer.PrintContainerLogsWithBanners(*dockerManager, ctx, testRunningContainerId, log, testRunningContainerDescription)
+
 	var testStatusRetrievalError error
 	switch kurtosisApiExitCode {
 	case exit_codes.TestCompletedInTimeoutExitCode:
 		testStatusRetrievalError = nil
-		// TODO this is in a really crappy spot; move it
-		banner_printer.PrintContainerLogsWithBanners(*dockerManager, ctx, testRunningContainerId, log, testRunningContainerDescription)
 	case exit_codes.OutOfOrderTestStatusExitCode:
 		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container received an out-of-order " +
 			"test execution status update; this is a Kurtosis code bug")
@@ -260,8 +263,6 @@ func RunTest(
 	case exit_codes.NoTestSuiteRegisteredExitCode:
 		testStatusRetrievalError = stacktrace.NewError("The test suite failed to register itself with the " +
 			"Kurtosis API container; this is a bug with the test suite")
-		// TODO this is in a really crappy spot; move it
-		banner_printer.PrintContainerLogsWithBanners(*dockerManager, ctx, testRunningContainerId, log, testRunningContainerDescription)
 	case exit_codes.ShutdownSignalExitCode:
 		testStatusRetrievalError = stacktrace.NewError("The Kurtosis API container exited due to receiving " +
 			"a shutdown signal; if this is not expected, it's a Kurtosis bug")
@@ -274,7 +275,7 @@ func RunTest(
 		log.Error("An error occurred that prevented retrieval of the test completion status")
 		return false, testStatusRetrievalError
 	}
-	log.Info("The test suite container running the test exited within the hard test timeout")
+	log.Info("The test suite container running the test exited before the hard test timeout")
 
 	// The test suite container will already have stopped, so now we get the exit code
 	testSuiteExitCode, err := dockerManager.WaitForExit(