+++ core/initializer/auth/auth0_token_claims/auth0_token_claims.go
@@ -3,14 +3,11 @@
  * All Rights Reserved.
  */
 
-package auth0_authorizer
+package auth0_token_claims
 
 import (
-	"github.com/kurtosis-tech/kurtosis/initializer/access_controller/auth0_constants"
+	"github.com/kurtosis-tech/kurtosis/initializer/auth/auth0_constants"
 	"github.com/palantir/stacktrace"
-)
-
-const (
 )
 
 type Auth0TokenClaims struct {