+++ core/initializer/auth/test_mocks/mock_client_credentials_authorizer.go
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 2020 - present Kurtosis Technologies LLC.
+ * All Rights Reserved.
+ */
+
+package test_mocks
+
+import "github.com/palantir/stacktrace"
+
+type MockClientCredentialsAuthorizer struct{
+	throwErrorOnAuthorize bool
+	tokenToReturn string
+}
+
+func NewMockClientCredentialsAuthorizer(
+		throwErrorOnAuthorize bool,
+		tokenToReturn string) *MockClientCredentialsAuthorizer {
+	return &MockClientCredentialsAuthorizer{
+		throwErrorOnAuthorize: throwErrorOnAuthorize,
+		tokenToReturn: tokenToReturn,
+	}
+}
+
+func (m MockClientCredentialsAuthorizer) AuthorizeClientCredentials(clientId string, clientSecret string) (string, error) {
+	if m.throwErrorOnAuthorize {
+		return "", stacktrace.NewError("Test error on authorization, as requested")
+	}
+	return m.tokenToReturn, nil
+}
+
+